pull report loading code out of conversation.tsx, make the report appear in sentiment checks page as well
