+++ client/src/pages/dashboard/report_comment.tsx
@@ -1,0 +1,103 @@
+/** @jsx jsx */
+import { Box, Text, jsx } from "theme-ui"
+
+export type ReportComment = {
+  active: boolean
+  agree_count: number
+  conversation_id: string
+  count: number
+  created: string
+  disagree_count: number
+  is_meta: boolean
+  is_seed: boolean
+  lang: string | null
+  mod: number
+  pass_count: number
+  pid: number
+  quote_src_url: string | null
+  tid: number
+  tweet_id: string | null
+  txt: string
+  velocity: number
+}
+
+export const ReportCommentRow = ({
+  reportComment,
+  maxCount,
+}: {
+  reportComment: ReportComment
+  maxCount: number
+}) => {
+  const { agree_count, disagree_count, pass_count, tid, txt } = reportComment
+  const row = { display: "flex" }
+  const bar = { px: "1px", py: "2px", lineHeight: 1.2 }
+  const text = { display: "inline-block", fontSize: "0.88em", left: "4px" }
+
+  return (
+    <Box key={tid}>
+      <Box
+        sx={{
+          bg: "bgOffWhite",
+          border: "1px solid #e2ddd5",
+          borderRadius: "7px",
+          mb: [1],
+          pt: "12px",
+          pb: "10px",
+          pl: "20px",
+          pr: "15px",
+          display: "flex",
+        }}
+      >
+        <Text sx={{ margin: "auto", fontSize: "0.91em", lineHeight: 1.3, flex: 1, pr: [1] }}>
+          {txt}
+        </Text>
+        <Box sx={{ position: "relative", pl: [3] }}>
+          <Box sx={row}>
+            <Box sx={{ width: 70, ...text }}>Agree</Box>
+            <Box sx={{ width: 70 }}>
+              <Box
+                sx={{
+                  width: `${Math.ceil((agree_count / maxCount) * 100)}%`,
+                  bg: "#2fcc71",
+                  color: agree_count / maxCount < 0.2 ? "#222" : "#fff",
+                  ...bar,
+                }}
+              >
+                <Text sx={text}>{agree_count}</Text>
+              </Box>
+            </Box>
+          </Box>
+          <Box sx={row}>
+            <Box sx={{ width: 70, ...text }}>Disagree</Box>
+            <Box sx={{ width: 70 }}>
+              <Box
+                sx={{
+                  width: `${Math.ceil((disagree_count / maxCount) * 100)}%`,
+                  bg: "#e74b3c",
+                  color: disagree_count / maxCount < 0.2 ? "#222" : "#fff",
+                  ...bar,
+                }}
+              >
+                <Text sx={text}>{disagree_count}</Text>
+              </Box>
+            </Box>
+          </Box>
+          <Box sx={row}>
+            <Box sx={{ width: 70, ...text }}>Pass</Box>
+            <Box sx={{ width: 70 }}>
+              <Box
+                sx={{
+                  width: `${Math.ceil((pass_count / maxCount) * 100)}%`,
+                  bg: "#e6e6e6",
+                  ...bar,
+                }}
+              >
+                <Text sx={text}>{pass_count}</Text>
+              </Box>
+            </Box>
+          </Box>
+        </Box>
+      </Box>
+    </Box>
+  )
+}
